Externalize Location Form as reusable component
